requirements.txt
# Core ML and Scientific Computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Token Counting (Claude API approximation)
tiktoken>=0.5.0

# Environment Configuration
python-dotenv>=1.0.0
